Move TqdmWriteStream to supply.ts
